// src/types.ts
export type SignKeyECDSA = {
    /**
     * Signature type
     */
    type: "ec-dsa";

    /**
     * ECDSA private key
     */
    key: Uint8Array;
};

/**
 * Key to sign a transaction with ML-DSA
 */
export type SignKeyMLDSA = {
    /**
     * Signature type
     */
    type: "ml-dsa";

    /**
     * ML-DSA spec
     */
    spec: 44 | 65 | 87;

    /**
     * Public key
     */
    publicKey: Uint8Array;

    /**
     * Private key
     */
    secretKey: Uint8Array;
};

/**
 * Key to sign a transaction
 */
export type SignKey = SignKeyECDSA | SignKeyMLDSA;